Ndtdoanh

+++ identity-service/src/main/java/com/hacof/identity/services/RoleService.java
@@ -35,6 +35,11 @@ public class RoleService {
     private final UserMapper userMapper;
 
     public RoleResponse createRole(RoleCreateRequest request) {
+
+        if (roleRepository.existsByName(request.getName())) {
+            throw new AppException(ErrorCode.ROLE_EXISTED);
+        }
+
         var role = roleMapper.toRole(request);
 
         var permissionsIds =
@@ -128,6 +133,9 @@ public class RoleService {
     }
 
     public void deleteRole(Long roleId) {
+        if (!roleRepository.existsById(roleId)) {
+            throw new AppException(ErrorCode.ROLE_NOT_EXISTED);
+        }
         roleRepository.deleteById(roleId);
     }
 }
